Combine the manifest test to existing test_manifest.py file
