+++ tests/utils/test_manifest.py
@@ -604,3 +604,26 @@ def test_write_manifest_list(
         assert entry.file_sequence_number == 0 if format_version == 1 else 3
         assert entry.snapshot_id == 8744736658442914487
         assert entry.status == ManifestEntryStatus.ADDED
+
+
+@pytest.mark.parametrize(
+    "raw_file_format,expected_file_format",
+    [
+        ("avro", FileFormat("AVRO")),
+        ("AVRO", FileFormat("AVRO")),
+        ("parquet", FileFormat("PARQUET")),
+        ("PARQUET", FileFormat("PARQUET")),
+        ("orc", FileFormat("ORC")),
+        ("ORC", FileFormat("ORC")),
+        ("NOT_EXISTS", None),
+    ],
+)
+def test_file_format_case_insensitive(raw_file_format: str, expected_file_format: FileFormat) -> None:
+    if expected_file_format:
+        parsed_file_format = FileFormat(raw_file_format)
+        assert parsed_file_format == expected_file_format, (
+            f"File format {raw_file_format}: {parsed_file_format} != {expected_file_format}"
+        )
+    else:
+        with pytest.raises(ValueError):
+            _ = FileFormat(raw_file_format)
